requirements.txt
Django==2.0.2
mysqlclient==1.3.12
gunicorn==19.7.1
requests==2.18.4
django-jsonview==1.1.0
